requirements.txt
discord.py==2.3.2
gTTS==2.3.2
pandas==2.0.3
pynacl==1.5.0
openpyxl==3.1.2
pymongo==4.5.0
googletrans==4.0.0-rc1
aiofiles==23.2.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
tabulate==0.9.0
